+++ src/Toolkit/CHANGELOG.md
@@ -1,5 +1,10 @@
 # CHANGELOG
 
+## 3.1
+
+- [Toolkit] Add `avatar` recipe
+- [Toolkit] Add `dropdown` recipe
+
 ## 3.0.0
 
 - Minimum required Symfony version is now 7.4